+++ lib/Convos/Core.pm
@@ -30,6 +30,10 @@ my %CONVOS_MESSAGE = (
 
 =head1 ATTRIBUTES
 
+=head2 archive
+
+Holds a L<Convos::Archive::File> object.
+
 =head2 log
 
 Holds a L<Mojo::Log> object.
@@ -43,8 +47,10 @@ Holds a L<Mojo::Redis> object.
 =cut
 
 has log   => sub { Mojo::Log->new };
+has archive => sub { require Convos::Archive::File; Convos::Archive::File->new; };
+has log     => sub { Mojo::Log->new };
 has proxy => sub { Convos::Proxy->new };
-has redis => sub { Mojo::Redis->new };
+has redis   => sub { die 'redis connection required in constructor' };
 
 =head1 METHODS
 
@@ -99,51 +105,6 @@ sub reset {
   return $self;
 }
 
-=head2 send_convos_message
-
-  $self = $self->send_convos_message($login, $message);
-  $self = $self->send_convos_message($login, $message, $cb);
-
-Used to add a C<$message> to the user C<$login>.
-
-=cut
-
-sub send_convos_message {
-  my ($self, $login, $message, @cb) = @_;
-  my $time = time;
-
-  local $CONVOS_MESSAGE{message}   = $message;
-  local $CONVOS_MESSAGE{timestamp} = $time;
-  local $CONVOS_MESSAGE{uuid}      = Mojo::Util::md5_sum($time . $message);    # not really an uuid
-
-  $self->redis->zadd("user:$login:connection:convos:msg", $CONVOS_MESSAGE{timestamp}, j(\%CONVOS_MESSAGE), @cb);
-
-  $self;
-}
-
-=head2 start_convos_conversation
-
-  $self = $self->start_convos_conversation($login);
-
-Will add default messages to the convos conversation.
-
-=cut
-
-sub start_convos_conversation {
-  my ($self, $login) = @_;
-
-  for (
-    "Hi $login!",
-    "While convos is establishing a connection, you can try out the help command: Type '/help' in the input field in the bottom on the page and hit enter.",
-    "You can also use the <tab> key to autocomplete commands and nicks.",
-    )
-  {
-    $self->send_convos_message($login, $_);
-  }
-
-  return $self;
-}
-
 =head2 start
 
 Will fetch connection information from the database and try to connect to them.
@@ -187,7 +148,9 @@ sub _connection {
   my $conn = $self->{connections}{$args{login}}{$args{name}};
 
   unless ($conn) {
+    Scalar::Util::weaken($self);
     $conn = Convos::Core::Connection->new(redis => $self->redis, log => $self->log, %args);
+    $conn->on(save => sub { $_[1]->{message} and $_[1]->{timestamp} and $self->archive->save(@_); });
     $self->{connections}{$args{login}}{$args{name}} = $conn;
   }
 
@@ -223,12 +186,10 @@ sub _start_control_channel {
 =head2 add_connection
 
   $self->add_connection({
-    channels => [ '#foo', '#bar', '...' ],
     login => $str,
     name => $str,
     nick => $str,
     server => $str, # irc_server[:port]
-    tls => $bool,
   }, $callback);
 
 Add a new connection to redis. Will create a new connection id and
@@ -238,19 +199,14 @@ set all the keys in the %connection hash
 
 sub add_connection {
   my ($self, $input, $cb) = @_;
-  my $validation = $self->_validation($input, qw( password login name server tls ));
+  my $validation = $self->_validation($input, qw( login name nick password server username ));
 
   if ($validation->has_error) {
     $self->$cb($validation, undef);
     return $self;
   }
 
-  $validation->optional($_) for qw( nick user );
-  $validation->output->{nick} ||= $validation->output->{login};
-  $validation->output->{user} ||= $validation->output->{login};
-
   my ($login, $name) = $validation->param([qw( login name )]);
-  my @channels = $self->_parse_channels(delete $validation->input->{channels});
 
   warn "[core:$login] add ", _dumper($validation->output), "\n" if DEBUG;
   Scalar::Util::weaken($self);
@@ -272,7 +228,6 @@ sub add_connection {
         [sadd  => "connections",                  "$login:$name"],
         [sadd  => "user:$login:connections",      $name],
         [hmset => "user:$login:connection:$name", %{$validation->output}],
-        (map { [zadd => "user:$login:conversations", time, as_id $name, $_] } @channels),
         $delay->begin,
       );
     },
@@ -294,7 +249,6 @@ sub add_connection {
     name => $str,
     nick => $str,
     server => $str, # irc_server[:port]
-    tls => $bool,
   }, $callback);
 
 Update a connection's settings. This might issue a reconnect or issue
@@ -304,15 +258,12 @@ IRC commands to reflect the changes.
 
 sub update_connection {
   my ($self, $input, $cb) = @_;
-  my $validation = $self->_validation($input, qw( login name nick password server tls ));
+  my $validation = $self->_validation($input, qw( login name nick password server username ));
 
   if ($validation->has_error) {
     $self->$cb($validation, undef);
     return $self;
   }
-
-  $validation->optional('user');
-  $validation->output->{user} ||= $validation->output->{login};
 
   my ($login, $name) = $validation->param([qw( login name )]);
   my $conn  = Convos::Core::Connection->new(%{$validation->output});
@@ -340,7 +291,6 @@ sub update_connection {
     },
     sub {
       my ($delay, $current, $conversations) = @_;
-      my %existing_channels = map { $_, 1 } $conn->channels_from_conversations($conversations);
 
       $conn = $validation->output;    # get rid of the extra junk from Connection->new()
 
@@ -412,6 +362,52 @@ sub delete_connection {
       $self->$cb(undef);
     },
   );
+}
+
+=head2 delete_user
+
+  $self = $self->delete_user(
+            { login => $str },
+            sub { my ($self, $err) = @_; ... },
+          );
+
+This method will delete a user and all the conversations, connections, and
+related data. It will also stop all the connections.
+
+=cut
+
+sub delete_user {
+  my ($self, $input, $cb) = @_;
+  my $redis = $self->redis;
+  my $login = $input->{login};
+
+  Mojo::IOLoop->delay(
+    sub {
+      my ($delay) = @_;
+      $redis->smembers("user:$login:connections", $delay->begin);
+      $redis->keys("user:$login:*", $delay->begin);
+    },
+    sub {
+      my ($delay, $connections, $keys) = @_;
+
+      $redis->del(@$keys, $delay->begin) if @$keys;
+      $redis->del("user:$login", $delay->begin);
+      $redis->srem("users", $login, $delay->begin);
+
+      for my $name (@$connections) {
+        $redis->srem("connections", "$login:$name", $delay->begin);
+      }
+    },
+    sub {
+      my ($delay, @deleted) = @_;
+      my $conns = delete $self->{connections}{$login};    # stop all connections
+
+      $self->archive->flush($_) for values %$conns;
+      $self->$cb('');
+    },
+  );
+
+  return $self;
 }
 
 =head2 ctrl_stop
@@ -517,14 +513,6 @@ sub login {
   );
 }
 
-sub _parse_channels {
-  my $self = shift;
-  my $channels = shift || [];
-  my %dup;
-
-  sort grep { $_ !~ /^[#&]$/ and !$dup{$_}++ } map { /^[#&]/ ? $_ : "#$_" } map { split m/[\s,]+/ } @$channels;
-}
-
 sub _dumper {    # function
   Data::Dumper->new([@_])->Indent(0)->Sortkeys(1)->Terse(1)->Dump;
 }
@@ -543,11 +531,11 @@ sub _validation {
 
   for my $k (@names) {
     if    ($k eq 'password') { $validation->optional('password') }
+    elsif ($k eq 'username') { $validation->optional('username') }
     elsif ($k eq 'login')    { $validation->required('login')->size(3, 30) }
     elsif ($k eq 'name')     { $validation->required('name')->like(qr{^[-a-z0-9]+$}) }    # network name
     elsif ($k eq 'nick')     { $validation->required('nick')->size(1, 30) }
     elsif ($k eq 'server') { $validation->required('server')->like($Convos::Core::Util::SERVER_NAME_RE) }
-    elsif ($k eq 'tls')    { $validation->required('tls')->in(0, 1) }
     else                   { $validation->required($k) }
   }