+++ t/commands.t
@@ -32,7 +32,7 @@ $t->websocket_ok('/socket');
 }
 
 {
-  $t->send_ok(msg('/help asdasd'));
+  $t->send_ok(msg('/HeLp asdasd'));
   $dom->parse($t->message_ok->message->[1]);
   ok $dom->at('li.help[data-target=""] dl'), 'HELP';
 }
@@ -76,8 +76,8 @@ for my $cmd (qw/ t topic /) {
 }
 
 {
-  $t->send_ok(msg('/list whatvereasdja n '));
-  is ws(), 'abc-123 LIST', 'abc-123 LIST';
+  $t->send_ok(msg('/list #chnl '));
+  is ws(), 'abc-123 LIST #chnl', 'abc-123 LIST';
 
   $t->send_ok(msg('/me is too cool :) '));
   is ws(), "abc-123 PRIVMSG #convos :\x{1}ACTION is too cool :)\x{1}", '/me is too cool :)';
@@ -87,6 +87,9 @@ for my $cmd (qw/ t topic /) {
 
   $t->send_ok(msg('/mode +o batman '));
   is ws(), 'abc-123 MODE +o batman', 'abc-123 MODE +o batman';
+
+  $t->send_ok(msg('/kick batman'));
+  is ws(), 'abc-123 KICK #convos batman', 'abc-123 KICK #convos batman';
 
   $t->send_ok(msg('/names #foo '));
   is ws(), 'abc-123 NAMES #foo', 'abc-123 NAMES #foo';